Added envelope cost, performance, and lifetime data
The EIA Annual Energy Outlook does not include data on the baseline cost, performance, or lifetime of components of the building envelope - roof, walls, and floor/foundation - that provide thermal, air, and moisture control for the building. These data were obtained from a range of sources and, because they are reported by climate zone, are added to the input database after the data are moved to a climate zone basis by the 'final_mseg_converter' module. These data were recorded in a separate database to clearly distinguish them from the data obtained from the AEO and to ensure that they were not lost within a module.

The envelope component cost data are originally reported in whatever units are common for that component, e.g., per square foot glazing for windows, thus an additional step was added to convert the cost data to a uniform per square foot floor area basis using a cost conversion table. These cost conversions are a function of building type since some conversions, such as window/wall ratios, will vary significantly between building types. This additional step ensures that conversion ratios are not embedded in the reported costs and that the costs on a per square foot floor area basis will be consistent with any updates to the conversion ratios or costs.

The cost units output by the residential and commercial building cost, performance, and lifetime data processing modules ('mseg_techdata' and 'com_mseg_tech') were modified to be uniformly consistent in the form 'YYYY$/ft^2 floor', where YYYY represents the year for which the U.S. dollar value is reported.

The 'mseg_techdata' module was updated so that it no longer adds temporary envelope data for residential buildings, thus enabling the 'final_mseg_converter' module to add cost, performance, and lifetime data for all relevant envelope components for all building types.

All of the tests were updated and expanded to cover the new and modified content of the modules changed in this commit.


Former-commit-id: 42e16e6fe6dcc9cb7241add0c04fb4e42304d9f0