Add supplemental cost, performance, and lifetime database
Supplemental cost, performance, and lifetime data have been developed
for baseline envelope technologies from a variety of sources, each of
which is noted in the file.

Additional data may eventually be added for Miscellaneous Electric
Loads (MELs) baseline technologies; a placeholder key for MELs is
therefore included in the database.

Closes #128.


Former-commit-id: e5b863d4d8e6bfa8ef1b712d268368febe05baf1