Fixed operation of dict_check()
dict_check() is a function originally developed as an analogue to the test method assertDictEqual, enabling the comparison of two dicts for equality but with an allowance for numbers that are slightly different due to floating point errors. The function was not causing test failures in several possible conditions that should have triggered failures. In each condition that should have failed, one of the two dicts: 1) had higher level keys that were not matching (e.g., 'apple' and 'pear', instead of 'first' and 'second') so long as the values for those keys were the same anyway, 2) was completely empty (i.e., {}), or 3) had keys that were entirely missing (e.g., only 'apple', instead of 'apple' and 'pear').

Case 1 was fixed by adding an explicit comparison of the keys each time the function is called.

Cases 2 and 3 were occurring a result of the way that zip() works, where any values that do not appear in both of the iterables being combined are dropped. As a result, with case 3, zip() would result in the test loop only comparing what was present in both dicts, and with case 2, nothing would be tested at all. Both cases 2 and 3 were fixed by using the more inclusive itertools.zip_longest(), which includes all of the non-matched items and provides a substitute value to maintain the ability to loop over all of the iterators provided.

These changes are related to issue #111, but do not fully resolve it since the original version of dict_check() appears on other active development branches.


Former-commit-id: b30afcb96611efdfea64f1154495b1d00fc86b24