Cleaned up measure attributes/markets, package handling, writing outputs
This commit includes the following changes:

1)	Measure attribute changes:
- A ‘remove’ attribute is included to flag measures with insufficient
source data (such measures will not be executed in the analysis engine).
- The ‘markets’ attribute has been changed to an OrderedDict for easier
readability in an output JSON.

2)	Changes to ‘fill_mkts’ function (retrieves measure market
microsegments):
- Missing consumer choice parameter cases now default to parameters
used for heating technologies.
- More general definition of secondary microsegments in the
‘partition_microsegment’ sub-function (secondary microsegments now do
not have to be tied to a lighting primary microsegment.)
- Prints message for user upon completing updates for each measure.

3)	Updated handling of measure packages:
- Measure packages now accept updated Measure objects, rather than
dicts containing updated individual measure attributes.
- A user receives a warning message when no updated individual measures
are available to package or available measures are flagged as invalid
(their ‘remove’ attribute is True).
- New ‘add_uncovered_pkgupdates’ function checks for packages that
would be affected by the updates to individual measures but are not
flagged for updating by the user; these packages are added to the
user-defined list of packages to update.
- A message is printed to notify the user of a successful package
update.

4)	 Updated output data organization and writing:
- Updated measure data are split into two groups via a new
‘split_clean_data’ routine: 1) detailed data on all market
microsegments that contribute to a measure’s overall market, which are
needed for future measure competition in the analysis engine, and 2)
summary-level data covering general measure characteristics (e.g.,
uncompeted energy/cost/carbon markets, applicable climate zones,
building types, end uses, assumed cost/performance/lifetime, etc.)
- Updated measure competition data are written to JSONs compressed
using the ‘bz2’ package; updated measure summary data are added to a
JSON of existing measure summaries.

Adjustments and additions to the tests in ‘measures_prep_test.py’
generally accommodate the above changes.


Former-commit-id: a0e598da9f45dd25c6e38a040e1188ca74a3bd4c