Improved performance re-baselining, added numpy conversion for JSON
The approach used to re-baseline measure performance over time is now
robust to cases where anchor year baseline performance is zero.

A new sub-class of the json module’s JSONEncoder has also been added
that converts numpy arrays to lists (numpy arrays are not JSON
serializable). Subsequent commits may explore the use of lists instead
of numpy arrays throughout the routine so that this conversion is not
needed.


Former-commit-id: 90be121e1bf7eca9d5c6afeda8c13b86a62f67f2