Adjusted data loading; fixed competition bug; added years on market
Market competition data for active measures are now loaded from pickled
and gzipped files, as pickle.load is faster than json.load for this
application.

Measure master markets data are still loaded from json files, but an
additional step is required to convert lists in these data to numpy
arrays (lists cannot be serialized using json). For this, a
‘convert_to_numpy’ function has been added and tested.

A bug has been fixed in measure competition where in a given year a
measure was being competed with other measures that were not on the
market. (This bug had previously been fixed for competition cases
without uncertainty in markets, but had not been fixed for cases where
uncertainty was present in markets.)

Finally, each measure’s range of years on the market is now directly
specified in a new ‘yrs_on_mkt’ attribute that is referenced during
competition routines. Tests in ‘run_test.py’ have been adjusted to
account for this new measure attribute.


Former-commit-id: 579c1ceda151f8c72410db9c676afe26b8b79d57