Improved efficiency of cost conversions that depend on building type
When the conversion of a measure’s cost to baseline cost units is
dependent upon building type, this conversion now occurs only once for
each building type in the measure’s applicable baseline market (e.g.,
it is not repeated for all the market microsegments within a given
building type).

In such cases, the printed message notifying users of a cost conversion
has been adjusted to provide information about which building type the
initial measure cost was converted for.


Former-commit-id: 84a75f2b5ca506b835f8d8d37d6f011fed5183af