Fix reference target, auto-number footnotes
To allow for references elsewhere in the document to point to a specific set of bullets, a section target link had been added. Because a section target requires new lines before and after, it disrupted the formatting of the nested list of bullets where it had been inserted. This section target (.. _section-target:) was replaced with an inline text target (inline `text target`_) and the reference text was updated to standard reference format (_`text target`).

Footnote references were changed from explicit references (e.g., [#f6]) to auto-numbered references (i.e., [#]) that are assigned numbers based on the order in which they appear in the markup when the documentation is built.


Former-commit-id: 36de68d0299a8303f634fe92db8e603333f0662f