Add text for the new ECM tutorial
Add text outlining the process for creating a new ECM and providing examples of the content for each field in an ECM. Modify the analysis approach section to provide references to specific language for the ECM authoring discussion. Also add a new folder with example files and include the example for the ECM described in the tutorial.


Former-commit-id: 7921dc1fe0bc7b3e503d4454eda24778374f9a77