Update tutorials
Update the new ECM tutorial example with corrections and improved explanations of some ECM features. Also update the example JSON to correspond with the changes in the tutorial. Add the second example ECM to the tutorials to illustrate some ECM definition features that do not appear in the first example ECM and add the corresponding reference JSON. Add the remaining tutorials: outlining the steps for preparing the ECMs and calculating the uncompeted energy, CO2, cost, and financial results; describing how to generate competed results; and explaining how to plot those results, and inspect and interpret the plots. Make tweaks, edits, and updates to all of the tutorials for clarity and accuracy. Closes #138.


Former-commit-id: f393a84f6ac01d8a9730dacefbf4db01bfcdbd69