Added check for rjson, adjusted y axis limits, fixed legend
The ‘plots.R’ routine now checks for an existing ‘rjson’ installation
on the user’s computer and installs this package if it is not already
installed.

The routine now also sets plot y axis limits based on min. and max.
values across the full set of data to be plotted, not just a subset of
these data as before.

Finally, an error in the line colors shown in the competed scenario
plot legend was fixed.


Former-commit-id: 11c27df83672fb9a8a01e57b3cc8d5218bbf2665