Revised approach to adjusting measure savings for competition
The approach used to scale down measure savings following competition
with other measures has been revised. Where previously this routine
scaled down the savings delta between a measure’s ‘baseline’ and
‘efficient’ energy/carbon/cost totals to reflect competition effects,
the routine now scales down the ‘baseline’ and ‘efficient’
energy/carbon/cost totals themselves, indirectly scaling down the
savings delta between them in the process.

While making the change described above, a fault was discovered in how
we account for the effects of competition on secondary microsegments.
Specifically, measure competition was affecting secondary microsegment
stock totals, even though secondary microsegment stock should not be
affected by a measure or its interactions with other measures. This
fault has been fixed.

A second fault was discovered in the method used to weigh the competed
market share calculated for a measure in the current year against the
market share calculated for all previous years (see
‘compete_adjustment’ function).  The competed market share weight for
the current year was calculated as the ratio of all competed and
captured stock in a given market microsegment to all captured stock,
when it should have been the ratio of all competed stock to all stock.
This fault has also been fixed.

Several of the test output values in the ‘ResCompeteTest’ and
‘ComCompeteTest’ tests of ‘run_test.py’ were adjusted to reflect the
above changes.  Additionally, an adjustment to the ‘dict_check’
function in ‘run_test.py’ was needed to ensure that the function
properly inspected key values formatted as numpy arrays. This
adjustment necessitated a change in the formatting of some output
values in the ‘PrioritizationMetricsTest’ test.


Former-commit-id: 69ebe800e5bac874b1ddee3fb758c662d40104fc