Determine ECM 'technology_type' attribute automatically
An ECM’s ‘technology_type’ attribute is now automatically specified
during ‘ecm_prep.py,’ based on the ECM’s ‘technology’ attribute.
Specifically, the ECM’s ‘technology’ value is checked against a
pre-determined list of technology names pertaining to the ‘demand’ side
of energy services; if the ‘technology’ value is in this list, the
ECM’s ‘technology_type’ is set to ‘demand’; otherwise, the ECM’s
‘technology_type’ is set to ‘supply’.

All instances of the ‘technology_type’ attribute have been removed in
test ECMs in ‘ecm_prep_test.py’ and in the ECM definitions already
posted to the Github repository.

Additionally, an output in the ‘FillParametersTest’ was amended to
reflect default handling of an ‘all’ ‘technology’ value as of the
‘supply’ ‘technology_type’ (previously it had been tested with a
‘demand’ ‘technology_type’).

Closes #150


Former-commit-id: 9d8093d5144940ce49381b69c184020001de0a24