Update package ECM sub-section
Update the sub-section with instructions on how to define package ECMs to reflect changes made to package ECMs that allow for the inclusion of cost reductions or performance improvements associated with grouping multiple ECMs into a package.

For the energy savings increase value, include a note that the value can exceed 1, indicating a savings increase of greater than 100%, but suggest that care should be taken to document such significant savings increases.


Former-commit-id: 8ff91f97db373e79f0bd2541df742691da634dc9