Fixed bug in windows market definition
In cases where an ECM applied to both windows conduction and windows
solar baseline markets, the data for those markets was supposed to be
combined in ‘ecm_prep.py’ such that windows would be competed as a
single unit with both conduction and solar energy components in
‘run.py’. A small error that had omitted the ‘windows solar’ data from
these paired markets was fixed.

Additionally, the market entry year was corrected in one of the
previously posted ECMs, and the length of package ECM names was
restricted for plotting purposes.


Former-commit-id: 871fa186373fa7280170e8cda4eb10b7add4cb14