Update Excel output method
Replace the R xlsx package with the WriteXLS package to output Excel XLSX formatted files with summary results. This change removes the Java dependence from the R code. Close #159.

Also add text with the total cost-effective energy savings to the top left corner of each of the financial metrics plots.


Former-commit-id: facb5c0a399340d57961638bfdcdeba0266bb44d