Update results figure information
Add information and examples to the tutorials about the two new results figures added to plots.R in f741ff83811f08f77d0eec14df68bb2215e2390a [formerly 1d149acc63d6877d33b15a0fdd7488ffdcaca859] (and a few earlier commits). Also move the Social Cost of Carbon URL reference to conf.py (and thus to the header of every section) so that it is common to all sections of the documentation.


Former-commit-id: e075176301c53563333758d7c140f105f3c1aa25