Update Windows quick start instructions
In 816577b09715c7d274b82076ff96e0869daa582a [formerly facb5c0a399340d57961638bfdcdeba0266bb44d], the method for producing the xlsx-formatted tabular outputs from plots.R was revised to eliminate the dependence on Java (addressing issue #159), but the new method requires Perl, which is not installed by default on Windows. Update the Windows section of the Setup Guide with instructions on installing Perl and verifying that the installation is recognized by R.

Further expand the Windows-specific setup instructions to improve clarity and avoid referring Windows users to the Mac instructions.


Former-commit-id: 2ee6fd3d05f96dbb50574793c4f875016225bfd3