Modified update message for packages, added deepcopy command
Messages about the initiation and completion of package ECM data
preparation in ‘ecm_prep.py’ are now printed to the same line of the
console, as was already the case for individual ECMs.

Additionally, a deepcopy of the individual ECM objects that contribute
to a package has been added back into the code; the removal of this
operation in the last commit possibly caused a Travis CI build fail for
Python 3.4.


Former-commit-id: 8a7088dbf5be572da4c5887178b9382fa28b681c