Improve documentation of figures
Simplify the documentation of the results figures by: reducing the extent of any repeated text between the figure captions and body text, adding sub-sections with headings for each figure type, ordering the presentation of the figures to match their order in a folder (when sorted by name), and including the file names in the discussion of each figure type. Replace the figures themselves to capture revisions to the analysis approach. Close #176.


Former-commit-id: ab50906504bf05af5e55e816280134e919fb3a08