Minor adjustments to 'run.py,' 'plots.R,' and cost conversion data
In ‘run.py,’ financial metric arrays for ECMs with uncertain inputs are
now initialized with NoneType values; the previous approach of
initializing with zeros yielded an error when trying to update these
values to dict outputs from the ‘metric_update’ function. Test outputs
in ‘run_test.py’ were updated accordingly, as was a piece of the
‘dict_check’ function in ‘run_test.py’ that checks for equality of
numpy arrays.

In ‘plots.R,’ adjustments were made: to the scaling of end use labels
in the individual ECM plots; to the method used to determine min. and
max. y values in the aggregate savings plots; and to the distance of y
axis labels from the plot region in the cost effectiveness plots.

Finally, a small correction was made to one of the units listed in the
cost conversion data.


Former-commit-id: cc89b4506273124d578827434cc51addda922bb6