Minor adjustments to 'ecm_prep.py,' 'plots.R'
In ‘ecm_prep.py’:

- the code no longer attempts to find baseline cost, performance, and
lifetime data for secondary micro segments, and
- the baseline ‘infiltration’ microsegment has been flagged as having
an installed cost of zero that should not prompt removal of this
microsegment from further analysis.

In ‘plots.R’:

- ECMs with more than two applicable end uses are labeled with
‘Multiple’ end uses on the individual ECMs plots,
- very small negative aggregate savings values are forced to zero to
improve the appearance of y axis limits,
- the y axis in the cost effectiveness plot is configured such that
points do not overlap with the top 5 cost effective ECMs labeled in the
upper right corner of the plot,
- the retrieval of data for the cost effectiveness plot now
accommodates a case where only one ECM is being plotted, and
- cost effectiveness plots are not generated in cases where no points
are available to plot.


Former-commit-id: c1ba16fb7d91ed8b502702775853447d10bc75ed