Reintroduce tests in ecm_prep_test.py and run_test.py
The 'MarketUpdatesTest' and 'UpdateMeasuresTest' classes in 'ecm_prep_test.py' and the 'test_compete_res_dist' and 'test_compete_com_dist' functions in 'run_test.py' have been reintroduced.

Modifications were made to some ‘MarketUpdatesTest’ output values to reflect an updated calculation of comparable baseline technology lifetimes for each ECM, and an updated calculation of secondary energy effects for lighting ECMs. In reintroducing the tests, it was revealed that the latter calculation was omitting a site-source conversion factor; this factor has been added as an input to the relevant ‘find_scnd_overlp’ function.

Closes #160.


Former-commit-id: d5d72ed3392fffdb46544157ec5a82fda1d5ebac