Improve financial calculation robustness, other run.py changes
Financial metric calculations in ‘run.py’ rely on competed stock figures, yet certain ECMs may not show competed stock in a given year (e.g., in a technical potential case, an ECM that applies only to existing structures may not show competed stock for multiple years after market entry). In such cases, financial metrics are set to those calculated for the previous year, excepting the first year in the modeling time horizon when they are set to 999 (representing NaN).

Additionally, an adjustment was made to how an ECM with a sub-market scaling fraction is competed; the portion of the competed market that the ECM does not apply to is evenly split across the other competing ECMs (previously it had not been divided by the number of other ECMs). This change still needs to be tested (issue #163).

Finally, in a technical potential scenario, competed market shares for years in which a competing ECM enters the market are weighed at 100% relative to all previous year market shares, reflecting that the entire baseline market is up for grabs in a technical potential scenario upon ECM market entry. Across all adoption scenarios, competed market shares for years without a competing ECM on the market no longer contribute to market share weighting for subsequent years.

Market entry years were added to test ECMs in ‘ResCompeteTest’ and ‘ComCompeteTest’ in ‘run_test.py’ to accommodate the last change.


Former-commit-id: 97873af277ce822750d248def0edd89d3d4f471a