Add example package ECMs to 'web' folder
Two example package ECMs were added to the 'package_ecms.json' in the ./web/ecm_definitions folder. These example package ECMs are now also reflected in 'ecm_prep_web.json' and 'ecm_results_web.json'. The example package definitions and results are needed for the Topcoder Scout web UI development work.

In running these package ECMs, it was revealed that a recent adjustment to the 'run' module required that packages have a 'measure_type' attribute. This attribute was added to the MeasurePackage class in the 'ecm_prep' module, and an associated test involving that class was changed accordingly.


Former-commit-id: 625d452033940fcb3fcfc0a69ba6840cccdee9dc