Remove baseline technology data requirement for special ECM case
In cases with missing baseline technology cost, performance, or lifetime data where the user specifies the measure as an 'add-on' type AND specifies relative savings for energy performance units, set the baseline cost to zero and baseline lifetime to that of the measure to ensure that all subsequent stock and energy impact calculations will continue for that baseline segment. Note: this marks a special exception to the general rule that baseline market segments without complete unit-level cost, performance, and/or lifetime data will be removed from further analysis. The exception is needed to handle controls-focused ECMs that apply to baseline market segments with poor technology-level data – for example, residential vacancy sensors that reduce MELs energy use by turning off power draws to circuits when an occupant isn't home. The user will now be able to evaluate such measures given measure relative performance, incremental cost, and lifetime data.

Test ECMs have been added to ‘ecm_prep_test.py’ that represent this special case in both the residential and commercial sectors; another test ECM has been added where the special case is not met and an error should be yielded.

The ‘verboseprint’ runtime function was assigned a normal function definition in ‘ecm_prep.py’ to facilitate its use in testing.


Former-commit-id: 907d5af80289cb10d1234e78d724a1c096358109