Fixed bug in calculation of efficient stock turnover
In a max adoption potential scenario, it is possible for a certain portion of the previously captured efficient stock to flow to competed stock due to efficient measure replacement or retrofit in a given year. While the competed stock fraction reflected this flow, however, the previously captured efficient stock fraction did not.

This bug is fixed by adjusting the previously captured efficient stock fraction to reflect the flow from this stock to competed stock in certain years.

The test of the max adoption potential stock-and-flow calculations in ‘ecm_prep_test.py’ was adjusted to reflect this change.

Additionally, the default year for ECM savings and cost-effectiveness snapshots in ‘plots.R’ was changed from 2030 to 2040.


Former-commit-id: f3e131ebacd5684264f18f861f6172a50a0931c1