Fixed inaccuracy in supply- and demand-side savings balancing
The definition of overlapping energy use across supply- and demand-side heating/cooling ECMs was overly broad, potentially yielding inaccuracies in the balancing of ECM savings across the supply- and demand-side of heating and cooling (described further in issue #205.)

The issue was addressed in the ‘run.py’ routine as well as in the supporting data file and routine ‘htcl_totals.json’ and ‘htcl_totals.py’. All associated tests were updated as well.

Closes #205.

Additionally, a small adjustment was made to the stock-and-flow calculations in ‘ecm_prep.py’ concerning cases where an ECM enters the market several years after the start of the modeling time horizon. In such cases, baseline technologies will capture all ‘new’ stock added after the beginning of the horizon, until this ‘new’ baseline stock eventually turns over and is replaced by efficient stock. The existing calculation method had overestimated the amount of this ‘new’ baseline stock that was turning over in later years.


Former-commit-id: ba30951f346b952790ac9f2c0eb23091b5f204e7