Added handling of large cost values in residential competition calculation
The residential competition scheme in ‘run.py’ enters a weighted sum of an ECM’s annual incremental capital and operating costs into an exponential function, creating the potential for very large and/or inf values when the net of these costs is high. Net cost is now capped at 00 to address the issue.


Former-commit-id: 178ce11002771a9fffbed31b5d450588b86220a2