Added handling of large cost values in residential competition calculation
The residential competition scheme in ‘run.py’ enters a weighted sum of an ECM’s annual incremental capital and operating costs into an exponential function, creating the potential for very large and/or inf values when the net of these costs is high. Net cost is now capped at 00 to address the issue.


Former-commit-id: a459032637c61d81df21c2fb981efaeaa3f6b375