Update packaging algorithms
The method used to merge the energy, carbon, and energy/carbon cost impacts of measures in a package was improved such that the combined impacts of overlapping measures are greater than the impacts of these measures individually but less than their sum total impacts. Methods used to account for overlapping stock, lifetime, and market scaling information across measures in a package were also improved.

The MergeMeasuresandApplyBenefitsTest in ecm_prep_test.py was expanded to ensure correct handling of the following packaging cases in particular:

* Envelope components and HVAC equipment
* HVAC and controls
* Envelope components, HVAC equipment, and controls
* Multiple envelope components

Another minor improvement to plots.R extends annotated impacts in some plots to two years in the model time horizon (from 2040 to 2030 and 2050).


Former-commit-id: 9df386aff70c90575534baeef136202639c332f8