Add exogenous HP conversion rates
Users are now able to specify external rates of conversions from various fossil-based heating/water heating and electric resistance heating/water heating technologies to heat pumps, using the command line argument --exog_hp_rates. External conversion data are found in the new file ./supporting_data_convert_data/hp_convert_rates.json. The rates are used to either construct an increasing baseline market for measures that switch to HPs (e.g., as more and more fossil-based or electric resistance segments are switched away from), or to erode the baseline markets of overlapping measures that do not switch to HPs (e.g., efficient gas furnaces). The application of the rates in this way precludes the need to compete such measures; accordingly, the microsegment keys for measures that switch to HPs are tagged with “-FS” to preclude competition with any other measures in the mix that apply to the same baseline microsegments but do not switch from fossil fuel to an electric HP.

When a user specifies exogenous heat pump rates, they will be prompted to select whether the rates apply to both regular new/replacement decisions and early retrofits, or whether early retrofits should be assumed to switch 100% over to the heat pump measures.

Several improvements were made to the “partition_microsegment” function to support the stock-and-flow calculations under exogenous heat pump conversion rates and to restructure the code to accept a more general, measure-specific rate of diffusion that is applied in a similar fashion to the heat pump conversion rates, to be further tested/finalized in a future commit.

A new set of test measures was added to the “MarketUpdatesTest” in ecm_prep_test.py and tested under the maximum adoption potential case to ensure that heat pump measures (and other non-HP measures) are assessed correctly when exogenous rates are/are not specified. A few tweaks were also made to other existing tests to support minor changes to the “partition_microsegment” function that were introduced alongside the exogenous heat pump conversion rate calculations.


Former-commit-id: 65e5578260a54d7b8c2f1c800888c6824b734e8e