Changes needed to execute release tests
The flag for cost unit translation in ecm_prep.py was amended to address a bug uncovered when running commercial controls measures that span multiple end uses (heating, cooling, ventilation, lighting).

The commit also revises a part of run.py that links all stock turnover calculations for measures spanning multiple end use technologies that include heating and//or cooling to a single heating or (if unavailable) cooling technology in the measure definition. This part of the code was not properly handling non-heating/cooling technologies in the measure definition (for example, lighting and ventilation technologies in a controls measure definition that spans heating, cooling, ventilation and lighting end uses).


Former-commit-id: 7c627a2b67790a998de2193f2992b3338fc6b48c