Address anchor year issue for rel. perf. calculations
Measures defined with units of “relative performance (dynamic)” include an anchor year to use in updating relative performance over time to reflect changing baselines. Older measure definitions (e.g., “Prospective Commercial Integrated NGHP”) may use an anchor year that is before the start of the modeling time horizon, yielding a KeyError in ecm_prep. Add a check in ecm_prep for this issue and if present, set anchor year to the first year in the modeling time horizon to avoid the error.


Former-commit-id: 0a0b289d92a385f729314578f2810d48a157e9c7