Tweak stock cost calculations/reporting and measure definitions
Total baseline stock costs are now based on the measure-captured portion of the stock only (rather than the entire available stock) to allow direct comparison of unit-level costs between the baseline and measure technologies (these comparisons divide total baseline and measure stock costs by the total captured measure stock).

This change required reintroduction of baseline stock cost adjustments for competition in run.py (since the baseline stock costs now depend on the amount of stock captured by the measure, which varies with competition).

An error in stock units reporting for residential envelope measures was corrected.

Residential roof measure definitions were amended to correct issues with their treatment as “add-on” vs. “full service” measure types; markets for NVC/AVC/CCHP measures were amended to exclude replacement of geothermal, and CCHP cooling performance was inferred to be 2X that of the heating performance targeted in the MYPP (consistent with treatment of heating performance for the cooling-focused NVC/AVC targets in the MYPP).


Former-commit-id: 49bde5298f99b37c4b2d8540e2113187d609cdce