Update measure definitions; minor bug fixes
Includes the following updates to measure definitions:

* Remove heating portion of prospective dynamic windows savings (likely overstated and will be replaced with improved method of assessment in the future)
* Add ventilation portion of commercial HVAC EE+DF measures
* Expand prospective commercial controls 30% impact target to include ventilation
* Remove “Commercial 90.1 Chillers + Env.” package and supporting measure definitions (yielded slightly negative savings in some regions)
* Correct error in cost units for two commercial water heating measures

Testing the expanded commercial measure definitions (w/ ventilation included) exposed a gap in the handling of cost conversions for such measures in ecm_prep.py, which was corrected. Similarly, a gap in the run.py sub-routine that links the stock turnover calculations for the ventilation portion of such measures to those of their heating portions was exposed and corrected.

Finally, for measures with effectively (but not exactly) zero savings, the competition fractions calculated via the --report_cfs option in run.py sometimes fall outside of the 0 to 1 range; these were all set to zero to avoid confusion.


Former-commit-id: df8789027bfb6e096606e6798389d57bf3fb9efb