Minor bug fixes to cost conversions, competition calculations
Testing with commercial measure definitions (w/ventilation included) exposed a gap in the handling of cost conversions for such measures in ecm_prep.py, which is corrected. Similarly, a gap in run.py that links the stock turnover calculations for the ventilation portion of such measures to those of their heating portions was exposed and corrected.

Finally, for measures with effectively (but not exactly) zero savings, the competition fractions calculated via the --report_cfs option in run.py sometimes fall outside of the 0 to 1 range; these were all set to zero to avoid confusion.


Former-commit-id: 1dea73562ed724cb3d6605dc5b2fbafac2a114bf