Corrections to envelope cost packaging
The calculations that support the –pkg_env_costs option were updated to fix an issue with how the envelope costs were being normalized to a per unit basis and added to the per unit costs of associated HVAC equipment in the package. In particular, the normalizing stock units in those calculations had not been properly harmonized across the envelope and HVAC data.


Former-commit-id: 5bc66f0a1c53cd5102207e9cf5760b5b290ca0b9