Implement full fugitive emissions workflow
Adjustments were made to the structure of fugitive emissions data in the ecm_prep module and its handling in the competition calculations of the run module to facilitate full execution of fugitive emissions calculations and reporting in the files ./results/ecm_results.json and ./results/agg_results.json. Results are reported under the following variable names:

*Baseline Fugitive Methane (MMTons CO2e)
*Efficient Fugitive Methane (MMTons CO2e)
*Fugitive Methane Savings (MMTons CO2e)
*Baseline Fugitive Refrigerants (MMTons CO2e)
*Efficient Fugitive Refrigerants (MMTons CO2e)
*Fugitive Refrigerants Savings (MMTons CO2e)

Rebase fugitive emissions branch on decarb-develop and address conflicts.


Former-commit-id: e3e43397b2aa896162388ce4df2f13d4a107c2e5