Update microsegment handling and cost and stock turnover calculations
In ecm_prep, a small correction in how cooling microsegments were handled under exogenous FS rates was implemented. Small improvements were made to the stock turnover calculations in the partition_microsegment function. A correction in how ‘add-on’ measure costs were merged with HVAC equipment measures in a package was implemented.


Former-commit-id: 01ed2a6c0a5b08d6a2c9a47208209c3ce70065ff