Update microsegment handling and cost and stock turnover calculations
In ecm_prep, a small correction in how cooling microsegments were handled under exogenous FS rates was implemented. Small improvements were made to the stock turnover calculations in the partition_microsegment function. A correction in how ‘add-on’ measure costs were merged with HVAC equipment measures in a package was implemented.


Former-commit-id: 5d93ad9e00e4d046392dc1d638e59711089f527e