Breakout stock for GCAM reporting
Stock data are now broken out and reported in the same way as energy, carbon, and cost data, with the exception that no “savings” values are calculated for stock (savings is not a meaningful concept for this variable). The breakouts are reported under the “Baseline Stock” and “Measure Stock” variables with appropriate units attached.


Former-commit-id: e9a09b8ca208ce851621f2cedeffe23fcfdef3cd