Add GCAM post-processing to run w/ supporting files
The run routine was modified to include a --gcam_out user option that maps measure portfolio impacts to GCAM segments, reflects those impacts relative to a reference case, and writes out the result.

Additional changes were implemented to reflect code from commit b6c415a on decarb-develop that has yet to be integrated with the master branch from which gcam-2022 is derived.


Former-commit-id: a701963af47a934fecd06bf36b45fb12fdab1b62