Adjust TSV load and cost calculations
When calculating hourly measure baseline loads and load impacts, the hourly baseline shape(s) pulled for a given measure segment from tsv_load are adjusted to ensure consistency with the hourly baseline shape in the measure’s savings shape CSV; this adjusted baseline is then scaled in accordance with the relative impacts of the measure in the CSV to arrive at the efficient/measure load shape.

When calculating hourly measure electricity price scaling factors, we now assume that only 41% of the retail electricity price is subject to scaling based on historical allocation of fixed vs. variable/volumetric retail electricity costs in https://www.nrel.gov/docs/fy22osti/78224.pdf. We also check for cases where multiplication of an hourly scaling factor by the retail rate will result in a negative price and force such cases to zero.

Apply a small adjustment to ensure that end use names used to pull CSV data are consistent with what's in tsv_load (e.g., "drying" in the former vs. "clothes drying" in the latter).


Former-commit-id: 1d7edbe3ce79840a559ce51732d13f368cf54c26