Update refrigerant calculations, adjust formatting
Estimation of total baseline refrigerant emissions in the first year of the modeling time horizon was adjusted to account for the typical refrigerant GWP of the installed base (e.g., all equipment stock that was installed before that first year). This assumed installed base of refrigerants GWP is that of the first year available in the refrigerants database, generally 2020.

type() was replaced with isinstance() throughout ecm_prep and run to conform with the latest pycodestyle requirements. A few other minor formatting tweaks were added.

Unspecified commercial microsegments without stock data are now handled by the stock turnover calculations, which assume a typical lifetime of 10 years to drive replacement turnover for these cases.


Former-commit-id: 029a2ef6075cec541d0e472bc63863bcd507e19c