Link heating/cooling microsegment turnover/switching calculations
Measures that apply to both heating and cooling equipment microsegments will now see stock turnover and (where applicable) exogenous HP switching rates linked across those microsegments. The linked calculations are anchored on a single end use and technology from the measure’s applicable market. The default end use anchor is set to heating (in UsefulVars), while the anchor technology is determined uniquely for each measure by comparing the technologies in the measure’s 'technology' attribute to a candidate list of anchor technologies for each building type (residential or commercial) and end use; the list order sets the priority for which anchor technology is used when the measure applies to multiple candidate technologies in the list.

In ecm_prep.py, for measures with linked stock turnover dynamics for cooling with heating microsegments (e.g., HPs), competed cooling stock are ultimately set to zero to avoid double counting across the heating and cooling microsegments; therefore, use of those competed cooling stock data in run.py would make it appear as though there was no cooling stock turnover at all to drive the competition of cooling segments. Modifications to run.py ensure that, for those measures, competed heating (or other linked mseg) stock data that are not set to zero in ecm_prep.py are used.

To ensure the anchor end use and technology are always updated first in the preparation routine, given the dependence of linked microsegments on these updates, the iterable list of microsegments to update for each measure is reordered in fill_mkts so that microsegments that include the anchor end use and technology come first. The rank order of heating/cooling technologies to choose as anchors for such within-measure linkages when multiple such technologies are present in the measure definition is harmonized across ecm_prep and run.

A slight adjustment to run.py was needed to handle a new string that is now added to the measure competition information to signify the specific heating/cooling linkages that the measure represents, if any (see use of variable “tch_apnd”).

Add reporting of efficient stock that remains with baseline technology when applicable (previously only reported measure stock under the efficient key).

Ensure that a measure's baseline and efficient market microsegment breakouts are always adjusted to reflect measure competition when breakout values are not None or zero across all years.


Former-commit-id: 5dc05af70af0358a4a7125643af49b26662374bd