Update measures and tweak meas gen code
Add a ‘global_exit_flag’ attribute that flags measures for exclusion from the market when a minimum performance floor is imposed in a certain year via ‘floor_start’ user option. (If this attribute is missing or None, measure will not be subject to exclusion via that floor option.)

Key measure changes:

* Add minimum efficiency FS measures (ASHPs, WH, boilers, cooking, drying) to the measure set.
* Add specific secondary fossil heating + envelope packages to measure set.
* Remove added electrical infrastructure costs for dual fuel res. Measures.
* Add electrical infrastructure costs for 240V HPWH measure, remove added circuit cost for HPWH TS measure.
* Remove commercial roofs measure, which was not generating energy savings relative to baseline.
* Breakout envelope measure performance by IECC climate zone (vs. AIA) where relevant.
* Add RTAC to cooling option for existing measures that pair chillers with heating.
* Ensure BTO RDO measures are consistent with latest roadmap measure definitions, including separating out infiltration from walls.
* Adjust cooking measure efficiency to be consistent with EIA docs.

In measure code, add checks for integers for market entry/exit year, and ensure that zero values are not inadvertently set to None in converting the CSV data to JSON format.


Former-commit-id: 654aad6a923df4ddf62e4c86b9ccb08fdff23794