Account for sub-federal appliance regulations.
Create and process new input data (./scout/supporting_data/sub_fed/state_appl_regs.csv) to represent the effects of state-level or other sub-federal regulations on the use of certain appliances (e.g., fossil-based appliances) for certain market segments, beginning in certain years.

The restrictions are reflected in compete_measures() as adjustments to the years a competing measure is allowed on the market (when the measure’s market is 100% restricted) or as a downward adjustment to its market share (when the measure’s market is partially restricted).

Addresses issue #409.


Former-commit-id: 9c37c89123ca361c6379c448748bdf00f11a765d